# api/openapi.json
{
  "components": {
    "schemas": {
...
      "httpx.APIError": {
        "properties": {
          "code": {
            "description": "Code is one of the stable Code* constants; clients branch on it.",
            "enum": [
              "INVALID_INPUT",
              "UNAUTHORIZED",
              "INVALID_CREDENTIALS",
              "FORBIDDEN",
              "ACCOUNT_DISABLED",
              "NOT_FOUND",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "RATE_LIMITED",
              "UNAVAILABLE",
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "message": {
            "description": "Message is a human-readable English message.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "httpx.ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/httpx.APIError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "server.ICEConfig": {
        "properties": {
          "ice_servers": {
            "items": {
              "$ref": "#/components/schemas/server.IceServer"
            },
            "type": "array"
          }
        },
        "required": [
          "ice_servers"
        ],
        "type": "object"
      },
      "server.IceServer": {
        "properties": {
          "credential": {
            "description": "Credential is set for TURN only and valid for 12 hours.",
            "type": "string"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "username": {
            "description": "Username is set for TURN only.",
            "type": "string"
          }
        },
        "required": [
          "urls"
        ],
        "type": "object"
      },
      "server.Legal": {
        "properties": {
          "legal_version": {
            "type": "string"
          },
          "media_retention_days": {
            "description": "MediaRetentionDays is 0 when media is kept until deleted manually.",
            "type": "integer"
          },
          "operator_country": {
            "type": "string"
          },
          "operator_email": {
            "type": "string"
          },
          "operator_name": {
            "type": "string"
          },
          "project_notice": {
            "type": "string"
          },
          "session_expiry_days": {
            "type": "integer"
          },
          "stun_servers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "turn_servers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "legal_version",
          "media_retention_days",
          "operator_country",
          "operator_email",
          "operator_name",
          "project_notice",
          "session_expiry_days",
          "stun_servers",
          "turn_servers"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "METRICS_TOKEN, only for /api/metrics.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "HttpOnly session cookie issued by POST /api/auth/login or /api/auth/register. The name is __Host-mnema_session when PUBLIC_URL is https.",
        "in": "cookie",
        "name": "mnema_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST API of Mnema Talk, a self-hosted, invite-only chat and voice app for a single community.\nAuthentication is an HttpOnly session cookie set by login/register. State-changing requests (anything but GET/HEAD/OPTIONS) must carry a same-origin Origin header, otherwise 403.\nJSON bodies are decoded strictly: unknown fields are rejected with 400 and bodies are limited to 1 MiB. Errors share the ErrorResponse shape.\nEverything except the SPA assets lives under /api.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit"
    },
    "title": "Mnema Talk API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/docs": {
      "get": {
        "description": "HTML page with Swagger UI over /api/openapi.json. Without a session it redirects to the app (/) to sign in. Only mounted when API_DOCS_ENABLED=true.",
        "operationId": "getAPIDocs",
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The API reference page."
          },
          "302": {
            "description": "No session: redirect to /."
          }
        },
        "security": [],
        "summary": "API reference",
        "tags": [
          "System"
        ]
      }
    },
...
    "/api/legal": {
      "get": {
        "operationId": "getLegal",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/server.Legal"
                }
              }
            },
            "description": "Legal information."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [],
        "summary": "Operator details and privacy-policy facts",
        "tags": [
          "System"
        ]
      }
    },
...
    "/api/openapi.json": {
      "get": {
        "description": "Returns this OpenAPI 3.1 document to signed-in members. Only mounted when API_DOCS_ENABLED=true.",
        "operationId": "getOpenAPI",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {},
                  "type": "object"
                }
              }
            },
            "description": "The OpenAPI document."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "This API description",
        "tags": [
          "System"
        ]
      }
    },
...
    "/api/webrtc/config": {
      "get": {
        "description": "STUN servers and, when configured, a TURN relay with fresh time-limited credentials. The list is empty when neither is configured.",
        "operationId": "getWebrtcConfig",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/server.ICEConfig"
                }
              }
            },
            "description": "ICE configuration.",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "ICE servers for voice",
        "tags": [
          "Realtime"
        ]
      }
    },
    "/api/ws": {
      "get": {
        "description": "Upgrades to a WebSocket (requires a same-origin Origin header; a foreign origin is refused with 403 by the upgrade handler). OpenAPI cannot model WebSocket frames, so the protocol is described here.\n\nEvery frame in both directions is a JSON text message `{\"type\": string, \"payload\": any}`. At most a few events per second per connection are processed; excess client events are dropped.\n\nServer to client, on connect (to this connection only):\n- `presence_snapshot`: map of user id to live status for every online user.\n- `voice_snapshot`: map of voice channel id to a map of user id to User plus `joined_at`.\n- `voice_rooms`: `{started: {channelId: time}, now}`; the server clock lets clients count up.\n\nServer to client, broadcast to all connections:\n- `message_create`, `message_update`: Message.\n- `message_delete`: `{id, channel_id, parent_id}`.\n- `message_reaction`: `{message_id, reactions}`.\n- `channels_changed`: no payload; refetch GET /api/channels (sent after any category/channel/layout admin change).\n- `member_joined`: public User of a newly registered member.\n- `user_update`: public User after a profile, status or avatar change, or `{id, disabled}` when an admin disables or enables an account.\n- `user_stats`: `{user_id, voice_seconds}` after a voice stay ends.\n- `presence_update`: `{user_id, status}` where status is online, away, dnd, focus or offline.\n- `voice_state_update`: `{action: \"join\", channel_id, user, started_at}` or `{action: \"leave\", channel_id, user_id}`.\n- `voice_speaking`: `{channel_id, user_id, active}`.\n- `typing`: `{channel_id, user_id}` (to everyone except the typist).\n\nServer to client, targeted:\n- `read_state`: to the user's own sessions only; `{channel_id, last_read_at, unread_count, mention_count}`, `{channel_id, last_read_at, refresh: true}` or `{channel_id, notify_level}`.\n- `webrtc_media_state`: `{channel_id, user_id, screen, camera}` to voice-room members.\n- `webrtc_offer` (SDP offer), `webrtc_candidate` (ICE candidate): SFU signalling.\n- `voice_kicked`: `{channel_id}` when an admin removes the user from voice.\n- `pong`: `{t}` echoing a `ping`.\n\nClient to server:\n- `ping` `{t}`; `presence_idle` `{idle: bool}`.\n- `typing` `{channel_id}` (text channels only, rate-limited).\n- `voice_join` `{channel_id}`; `voice_leave`; `voice_speaking` `{active}`.\n- `webrtc_answer` (SDP answer), `webrtc_candidate` (ICE candidate), `webrtc_request_keyframe`.\n- `webrtc_subscribe` `{kind: \"screen\"|\"camera\", user_id, on}` or `{kind: \"camera\", all: true, on}`.\n- `webrtc_screenshare_stop`, `webrtc_camera_stop`.\n\nUnknown event types are ignored. The server closes the connection when the session is revoked or the account is disabled.",
        "operationId": "connectWebSocket",
        "parameters": [
          {
            "description": "Must be websocket.",
            "in": "header",
            "name": "Upgrade",
            "required": true,
            "schema": {
              "enum": [
                "websocket"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols; the connection is now a WebSocket."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "No valid session (UNAUTHORIZED): missing, expired or revoked cookie, or the account was disabled."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "FORBIDDEN: not allowed (admin required, not the author, wrong current password) or cross-origin request rejected by the CSRF check."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "RATE_LIMITED: too many requests.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the client may retry.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/httpx.ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR: sanitized server failure."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "WebSocket event stream",
        "tags": [
          "Realtime"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}